+++ src/cli.ts
@@ -41,6 +41,7 @@ import { a11yLintCommand } from "./commands/a11y-lint.js";
 import { flowCommand } from "./commands/flow.js";
 import { contentLintCommand } from "./commands/content-lint.js";
 import { vrCommand } from "./commands/vr.js";
+import { evidenceCommand } from "./commands/evidence.js";
 import { ingestFigmaDsCommand } from "./commands/ingest-figma-ds.js";
 import { synthesizeConventionsCommand } from "./commands/synthesize-conventions.js";
 
@@ -85,6 +86,7 @@ COMMANDS[a11yLintCommand.name] = a11yLintCommand;
 COMMANDS[flowCommand.name] = flowCommand;
 COMMANDS[contentLintCommand.name] = contentLintCommand;
 COMMANDS[vrCommand.name] = vrCommand;
+COMMANDS[evidenceCommand.name] = evidenceCommand;
 COMMANDS[ingestFigmaDsCommand.name] = ingestFigmaDsCommand;
 COMMANDS[synthesizeConventionsCommand.name] = synthesizeConventionsCommand;